Add features to play next song in queue
The private next function has been added to trigger after the current
song has finished playing. This should play the next song in the queue if there
is more than one song in the queue. The skip command works as a manual
trigger that stops the current song and plays the next song in queue.