Add feature to add group permission to config
Within the "perm" command and the "group" sub-command is the "add"
sub-command which should be used to apply command perms to a discord
role. It checks to see if the command exists and if the group already
contains the perm before adding it to the specific server's config file.

Permission checks have been added to the alexa module.