Add perm removal & Fix existing perm check
The remove sub-command can be used to remove a perm that has been
assigned to a group. It should first check to see if the command exists
and if the group already contains it.

The check to see if a group already has a permission originally grabbed
the context command rather than the specified command, which lead to the
check succeeding every time rather than failing when required.