Rework permission system to use SQLite database
This is a massive overhaul for the permission system that replaces the
per server ini configs for an SQLite database. Currently, the database
should contain a seperate table for groups and users, with the server
ID, user/group ID, and command permission in each. This should increase
speed and efficiency when dealing with many concurrent requests.

The check that is run on every command has been modified to implement
the changes in permission storage, making use of SQL querying to check
for the requred user/group perms for the user that runs the command.
Adding new perms has also been reworked, applying the proper checks and
SQL insertions.