Add group permission parenting
The perm helper has been updated to include the group parent table,
allowing groups to inherent permissions from other groups. This has been
created alongside the group parent subcommand, allowing parents to be
checked and set. Parent can only be added if group doesn't already have
the parent, and if the parent isn't already the child of the group.