Rework config file to use TOML rather than INI
The need for TOML arose when the barebones INI functions hampered the
ability to properly handle lists and various other features. TOML is
essentially an INI formatted config with functions from JSON, allowing
for a greater use of persistent settings storage without having to use a
database for everything.

Porting an existing INI config to TOML will not be provided due to the
bot still being in development stage. As the current config file should
still be small enough to handle by hand, all that is required is to set
existing headers to lowercase, and to wrap strings around double quotes.