Rework permssion system to use wildcards
Added the ability to use wildcards with permission assignment. Original
command assignment of "command" can still be used, as the same command
name cannot be used twice in the bot. These commands will be saved as
"Module.command" in the database. Command permissions can also be
specified with the module name, such as "Module.command". Wildcards can
be used in place of the command name such as "Module.*" to give the
target access to all commands in the module. A simple "*" gives
permission to all non-exclusive commands.

Existing database permissions cannot be used, and no feature will be
implemented to transfer over existing permissions to the new system as
the bot is still in the rapid development phase.
Any existing permission databases are assumed to only be used in
non-production testing environments.