Fix stream invalidation during music playback
Youtube_DL constantly errored during during stream playback for some
reason (literally the reason in the console). Set before_options to
reconnect in the case of this occuring, which is all the time. Also
silenced ffmpeg to not spam the console with pull function errors,
which are solved by the reconnection.