Fix multi server music player conflict
Changed music variables to be dictionaries, utilising server IDs as keys
to ensure that each server has its own individual queue and toggles.
Keys are dynamically added and removed using a function on join/leave,
with an additional function to display a message in the case that a user
tries to execute commands while the bot isn't in a voice channel.