Make looped song use YTDLSource class
Using the discord FFmpegPCMAudio function directly caused issues with
volume and not applying the settings required to prevent stream errors.
This caused the loop function to essentially be broken. A thread safe
coroutine was created within the _next (song) function toactivate the
async from_url function, as the _next function isn't able to beasynced
due to how functions called after playback is handled.