Add dynamic config settable prefix
A few additions have been made to allow prefixes to be customised. The
first run setup now includes the addition of asking the user to set a
prefix. This gets saved to the config to be retrieved later. The bot
object and command prefix is now created inside of the startup class
init, which also changes how the bot object must be fetched within the
SpaceCat module. This now gets the prefix from the config rather than
being hardcoded.