Optimise how servers are checked and added to db
Using INSERT OR IGNORE INTO ensures that no duplicate entries will ever
exist in the server_settings table. Doing the comparison to check if a
server is already in the database is handled quicker by fetching the
entire table's server_id column and comparing it to a list of servers
that the bot is in rather than one at a time SQL executions.