Seperate config setup to have finer control
Config setup of admin, prefix and invite have been seperated to have
finer control over what parts need to be setup. This ensures that in
the case that only one of the requirements are missing, only one of the
configurators needs to be called. This solves the issue of the last
merged branch breaking if prefix is not set in config.