Add test for custom alises using on_command_error
Custom aliases could be implemented using various different methods.
While this could be done using on_message, on_command_error has been
used instead so that it only does the appropriate checks if the message
starts with a valid prefix and fails to find an existing command. The
test currently checks for the "!test" command and executes the "!ping"
command instead.