Create a makefile to handle running & distributing
The makefile contains a variety of functions used to ease the build,
run and distribution processes. This was mainly designed to handle the
development process in a virtual environment. Self explanatory
instructions can be viewed by typing "make" in the root directory.

Manifest file included to include the license file in distribution,
with the gitignore file modified to exclude the dist folder in git
commits.