Add all data required for playlist music
The add subcommand of playlist employs the checks required to add a
song to the correct playlist in the correct order. The title, length and
url is fetched from the YTDL source. previous_song is fetched by
comparing the song ids and previous ids in a playlist to check which
song isn't being used as a previous song. The playlist id is fetched by
using the unique playlist name available on the server and fetching the
linked id.