Add custom ValueError exception for playlist get
ValueError has been used in place of a direct return none for _get_songs
and _get_playlist when a specified name does not exist in the database.
Each command that uses these functions have had the initial check
changed from TypeError to ValueError as the error type makes more sense.