Reimplement music status check to queue subcommand
The 'invoke_without_command' argument applied to the command group
command allows for the use of arguments in the parent command, though
it disables running the parent command at all. This makes
'if ctx.invoked_subcommand' redundant and requires the appropriate
checks to return back to each subcomand as they're treated as entirely
seperated commands.