Change how song time in queue switch is handled
EAFP principle has been replaced with LBYL in deciding whether to use
the pause song time or the start song time for the queue. This makes
processing times even rather than favouring the speed of when the song is
paused, as the try/catch would result in the exception being called in
most cases.