Add additional config creation checks
Setup now checks for both the config file and the API key to determine
if the apikey entry is missing despite the config file existing. This
should force the introduction to run. The introduction now also checks
to see if a config file already exists so it doesn't override everything
while trying to just add the API key.