Add custom exceptions for failed video fetching
The two self explanatory exceptions VideoTooLongError and
VideoUnavailableError are used in place of the vague ValueError used in
music download processing so that the error messages that get printed to
the chat can get offloaded from the processing function into the command
functions. This way, the playlist adding doesn't display error messages
while songs are being added to the queue.