Fix unavailable song checks when first song
A loop was applied when the first playlist song is to be instantly
played so that if the first x amount of songs in the playlist are
unavailable, it'll keep looping until it finds an available song to
play. Additional fixes were also done to ensure that the playlist will
not skip over the first song when a song is already playing.