Add subcommand help menus
The inputted names is now split into multiple individual strings. The
first string is checked as the top level command, with each string
thereafter checked as a lower level of subcommand. The same command
info embed is used for subcommands, with the parent commands inserted as
a sort of prefix to both the title and the usage sections.