Add command help page permission check
Users without permission to run the command are unable to view the
command's help page. While they are still able to view the help page
from users who are able to access the command, it serves as an indicator
that users shouldn't need to view pages they don't have access to
anyway.