Add default permission preset to config
A new category has been created in the config file, labed 'permissions'
to hold permission presets that can be applied to server roles. The
default role that is automatically generated is meant to hold
permissions that everyone can use. Permission removal should be able to
be implemented afterwords to revoke these.