Add permpreset check for proper cmd module handling
With the previous implementation, prefixing the command name with
any module allowed it to be added anyway, prefixed with the proper
module in the config. Checks are applied to ensure that only the proper
module can be prefixed, even though a prefix doesn't need to be added
in the case of adding command permissions. This same issue is present in
the 'perm' command implementation.