Add advanced_permission database entry with command
This new table key is used to house the server's permission mode.
False indicates that the server will use the bot's default config
permission assignment, which is necessary for providing basic user
functions without manual permissions. True indicates that the server
will now handle default (@everyone) permissions manually, requiring the
necessary use of manual assignment if users are to use any commands at
all.

The command assigned to handle this flip, 'perm advanced' is a
subcommand, which connects to the database and togggles the mode, telling
user of the changes that must be done because of this change.