Add perm advanced check and general tweaks
Config default permissions is now checked after both administrator and
database checks. The database is checked beforehand to determine if the
server is running advacned permission mode or not so it knows whether
to skip the config check. General tweaks have been made to streamline
the overall checks, as the function being executed on command call can
get very taxing.