Optimise how permission checks are done
Few tweaks were done to ensure that the perm check function, which is
called on every command execution, runs faster and uses less memory.
Sets were used rather than lists for performance reasons, allowing the
use of intersections as a means to compare sets of data rather than
iterating in a for loop.

Database now fetches all permission results
rather than the ones it needs. This may or may not be more memory
intensive, though it does speed up processing time on the python side.

A row_factory was used so that the data that is fetched from the
database aren't tuples nested within a list, but rather just a list.