Add an instance menu on launch
On launch of the discord bot, rather than being launched directly into
the bot, a menu pops up requesting the user to either select an
instance or to create a new instance. Instances are detected by checking
for a config file in each subdirectory of the data folder.

ASCII art has been added as a header for fun. :D