Fix various instance deletion based checks
Various changes through the instance module broke the destroy instance
menu. Checks were reconfigured and the proper function is called so that
it works just as before. The name of the instance has also been added to
the confirmation prompt so that the user is aware of which instance they
are deleting.