Split config checks properly to avoid deletion
If the API key were to be removed from a fully configured config file,
it'll get wiped on the next run because the program thinks you haven't
gone through the initial setup yet. The checks are now split into
checking if the config file exists and checking if the apikey key exists.