Use importlib to fetch embed icons
Using importlib resources allows non code files to be imported and
accessed like code imports. This change means that it's easier to access
with an enum rather than a dict. All embed icons have now been moved
into a folder called 'resources' in the spacecat directory which is
included when installing the bot.