Switch image icons in embed with emojis
To remove the reliance on external image resources, embed icons now make
use of existing emojis which load faster and reduce complexity. This
means that embeds should now use the title parameter rather than the
author parameter to support the use of emojis.

Side note: The emojis for the embeds currently do not scale or position
correctly, but that's a discord client problem.