Refactor EmbedStatus to use enums for embed colors
To change all grouped constants to use enums, embed colours have been
moved to an enum class named EmbedStatus. A few of the embed colour
variables have also been changed. Warn has been changed to fail to make
it more clear that it's being used for failed command inputs. To more
easily differentiate between the on and off for a toggle command, there is
now seperate colours for yes and no embeds. Changes implemented for
all embeds throughout the bot.