Tweak existing embed colour selection
Some embeds are using the incorrect colour, some because of the new NO
embed colour and others because of poor testing. YES is now used for
positive successful results such as adding a song to queue, and NO is
used for negative successful results such as removing a song from queue.
INFO is for simply informational embeds, fails for invalid command usage
and SPECIAL for special alerts, which is currently unused.