Add command to list linked channels
To list the channels linked by the linkchannels command, the
listlinkchannels command has been created to list all the linked
channels on the current server. The channels on the embed also double
as links to the channels themselves, so users can directly see which
channel is which in the case that there are duplicate channel names.