Add linked channel index and pages
To limit the amount of linked channels that show on one embed, the
listlinkchannels command now only displays the first 10 results, with
the rest of the links accessable by specifying a page number. Due to
this, the index of each link has been added to provide a reference point
as to what page they're on.