Remove run and upgrade scripts
As the upgrade scripts are not required to upgrade from the previous
major versions, they have been removed as anyone currently running this
would've probably upgraded their databases using the scripts by now.
The start scripts have been removed and instructions for running
in a virtual environment if necessary will be provided.