Extend line limit to 99 characters
A new limit has been implemented which allows for more leniency in terms
of characters per line. While the standard 79 characters is still
recommended for most cases, the new limit of 99 can be taken in effect
in the case that there is absolutely no elegant way to split a line.
This essentially serves as the new hard limit, enforced by Flake8, while
the original 79 is a soft limit guideline. Existing code has been
rearranged as such.