Attempt to get module through slash client
This seems to be a really roundabout way to do things, but unlike
discord.py's shortcut method of getting the cog of the command, the
slash command API doesn't have that sort of implementation yet. This is
yet to be tested and is currently only in place as other changes still
need to be made in order for testing to be done.