Modify the error handler to handle slash commands
While the slash command error handler makes use of the same decorator,
the method name itself and the error instances it goes by are different.
Most of the previous checks can be ignored as they are handled on the
discord client itself, though checking for user and bot permission is
still required.