Update to experimental discord.py v2 API
This update though still in the development phase allows for integrated
use of slash commands among other changes. It is also a necessary
change for the near future as the previous API will be depreciated by
discord and make the bot unusable.

The update will cause many breaking changes that will have to be fixed.