Add custom bot class that inherits commands.Bot
In order accomodate the changes to extension loading becoming async, the
function to call the extension loading must also be made async. The
easiest way to handle this was then to make use of the setup_hook method
overload as part of the Bot class.

A new class named 'SpaceCat' was then created that inherits the Bot
class to handle module loading. The existing SpaceCat class that
handles the core commands was then renamed to 'Core'.