Remove linkle module due to new discord feature
Discord added text chats for every voice chat, which aims to accomplish
the same goals as this module. Therefore, this module is now redundant.
While there are additional use cases (linking multiple channels to a
single voice channel) that can come out of this module, the upkeep just
isn't worth it.