Rework help command to work with slash commands
Due to the incompatibility between the way descriptions and parameters
are handled between the original command system and the slash command
system, method calls had to be changed around.

One of the major issues was that there wasn't a clear cut way to get
posix-like command signatures, so a protected field had to be accessed
and formatting done manually to replicate the original help command
system.