Separate server specific dicts into its own class
This is a cleaner way to separate server specific instance data. Rather
than having a dictionary for each option that contains a key of each
server, a new instance of ServerMusic is created instead which contains
all the options it needs for each server. This will eliminate the need
to generate keys for each option manually.