Add methods for start and stop in MusicPlayer
Methods now exist decoupled from the Alexa class that allow for non text
output versions of music playback. The current setup allows for multiple
command methods to invoke the same actions, as the previous method for
invoking commands is now unavailable. This also eliminates the need to
tweak individual keys and is instead handled all in one place.