Automate adding and playing songs
To eliminate duplicate code to do with checking whether the queue is
filled before choosing whether to play a song, the functionality is now
built into its own function. The "play" function is now repurposed to
instantly play a song, with a separate "add" and "add_multiple"
functions that adds to the queue and plays the song if the queue is
empty.

An enum has been added which these functions return so that appropriate
outputs can be created depending on the result.