Create Scheduler module to run events later
This module's purpose is to allow users to create events to be run later
or at set intervals. This can be something as simple as a message
reminder or user settings manipulation.

This currently implements the base "Scheduler" class to hold commands
and listeners, as well as a simple "Reminder" class to hold messages to
be sent later.