Add 'remindme' command with timestamp conversion
To test the user facing interaction with the reminder system, a command
has been created with all the arguments a user could need to specify
when the reminder should be sent out. Each time based argument is
optional.

A static method to convert these intervals to a seconds based timestamp
has been added to be used by this command.