Add reminder loop to send out reminder calls
The reminder loop makes use of a checking system that finds the earliest
reminder to be dispatched and waits until then. Once that reminder is
dispatched, it is handled by a listener to have a message sent out, and
then subsequently refreshes the checker. The checker is also refreshed
upon certain commands being called that manipulate the reminder system.

This was done as an alternative to a task loop system in which checks
are done in set intervals to eliminate excess checks. A fixed time loop
would require the database to be checked on every loop, and depending on
the set intervals could prevent smaller timers from being called on
time.