Replace schedule library with RepeatJob
RepeatJob is a custom implementation of the repeating scheduler, a class
that takes in an event with a repeating value and runs its own isolated
loop to calculate when to next run. Like with the single firing events,
this does an async sleep using the interval to continue when the event
is to be fired. This will hopefully be more performant than the previous
implementation that checked every second.