Refactor event "Args" or "Function" to "Action"
Args and Function are technical names that are already prominently used
in software development, making it very confusing when the name is
meant to reflect a user facing functionality. "Action" is a better name
used to describe something that an Event is to do, and existing
references to do with event functionality should use this instead.