Fix adding action repos to service dictionary
The fix implements a (janky?) solution making use of __orig_bases__,
which allows it to get the concrete type that replaces the generic. This
now properly saves the type name of the action into the dictionary. This
also meant that the get_name functions had to be converted to class
methods.