Fix event action dispatch logic
The event dispatcher was still sending the event object and using the
old processing system for getting the event args. This new system
instead sends the action objects and fetches the arguments as each
individual variable. The proces sof doing so has also been split between
the RepeatJob and EventServer to handle the two parts of the job.