Change one time events to not auto delete
Rather than auto deleting, it removes the dispatch time and set a last
run time in order for it to be reusable at a later date without having
to set an interval. This makes one time events more reusable, especially
if you have a bunch of actions set that take a lot of effort to
configure.