Import wavelink dependency
Rather than make use of discord.py's built in player implementation,
this feature test will attempt to make use of wavelink, a python
implementation of the lavalink centralised player. It contains a lot of
features that can outright replace some of the custom implementation in
the alexa module, but it should still be better to make use of
interfaces to ensure that any future changes in player is easy to swap
out.