Add concrete implementation for Wavelink MusicPlayer
The new methods from the MusicPlayer abstract have been implemented as
best as possible. The queue system makes use of python's built in deque
system rather than with a list, as deque is more performant when doing
queries on the start and end. While queries in the middle will have to
be performed from time to time, it isn't a common operation.