Separate query and youtube playlist fetch
The way as provided by wavelink seemed like a bit of a hack and my
implementation hijacked that hacky way. Instead, playlist data fetching
has been separated, with a check to see if the url is a playlist url
before selecting which type of classmethod to use to fetch an audio
source.