Change AudioSource to default to YouTube Music
By defaulting to YouTube Music, it automatically excludes non music
videos, music videos, etc. This does mean that more obscure results may
have to be manually done by the user through YouTube, but this is a
primarily music focused module so you'll have to make do with whatever
is on the YouTube Music library.