Add utility to provide more spotify playlist data
Since the wavelink spotify extension doesn't provide concrete
information when given a spotify playlist, this new utility adds a
custom implementation of the spotify search to provide this. A new
datatype of SpotifyPlaylist now contains the playlist name, alongside
the track data on the spotify database. While this won't accurately
match up with the youtube database in all cases, it means that there's
at least something displayed.