Add clickable queue songs and reduce display limit to 5
Clickable queue songs is a useful addition to the end user to see where
the song was fetched from, and allows them to view it on that platform.
Due to links adding onto the embed text limit, the queue had to be
reduced to 5 songs per page in order to not hit the limit.