Change track skipping to work with prev
Rather than the skip command stopping the player and waiting for the
auto next track system to play the next song on its own, it now calls
the next function directly. This allows for the prev command to function
by calling the previous function to do the opposite, without the auto
track player getting in the way.