Add song command to display song specific info
While this doesn't really add much info other than what you would find
in the queue command, it could be useful as a source of additional data
in the future. Currently, it contains the additional data of where it
fetched the data from, as well as a link if it's from a YouTube or
Spotify album/playlist.