Fix rapid skip through queue two command calls in succession
When one or more users call the skip command in quick succession (around
0.1 seconds apart), the queue will be rapidly skipped through until
reaching the end. This is caused by a boolean flag that is immediately
flipped after processing the first command, causing the second skip to
infinitely trigger a song end event. By adding a slight 2 second delay
on song start, it ensures that both skips are processed before the flag
is reflipped to allow auto playing again.