Fix rapid skip without using delay
The previous fix attempt made use of a delay to process all skip
commands within that time. Unfortunately, it comes with another bug
where if you added a song under 2 seconds to the queue and then skipped
to it, it would prevent the queue from continuing to auto play songs.
This new implementation just stops the song on using a /skip or /prev
command, making use of a direction flag to check which direction to go.
Therefore, it should only be able to skip if a song is currently playing
rather than allowing the user to skip while in between songs.