Add regex to remove username from spotify link
Spotify playlist urls that included the username couldn't be parsed
properly. This is probably due to the username in the url being
exclusive to users viewing their playlist on their own account, with
other people being redirected to a link without the username. The API
isn't capable of doing this automatically so a simple regex is here to
fix it instead.