Add a generic paginator template
This template should allow users to set a base embed, and append a
paginated list to it. It should be able to display a list of items that
the user has set and how many entries per page. Pages should be able to
be navigated by a set of UI buttons.