Add Throwing class to store data from running throw command
This is a new system that should allow users to "catch" a throw
activated by the throw command. To do this, data pertaining to the
thrower and target must be saved in some form or another. It should also
hold the time period in which the ability to catch will expire.