Add class for RPS button and game
The game class stores data for the game of RPS instead, which holds info
such as who is playing the game and what actions have been taken,
alongside methods to play the game and determine the winner.

The RPS button is a subclass of Button that stores additional action
data, which is then utilised to determine what move has been made so
that it can be passed onto a stored RPS game instance.