Fix timeout on DefaultView by using its own send method
This custom send method bypasses the need to manually send and save the
message as an attribute and instead handles it on its own when you call
the send method through the view. Buttons are now automatically timed
out based on a default 5 minute period, but can be modified by any
view that requires a custom value.