Rework RepeatJob class into EventScheduler
This is a big scheduler overhaul that changes how repeating events are
handled. Rather than wrapping events in a new class that handles each
event in isolation, the new EventScheduler class now works in a broader
scope, whereby events are simply accepted by a schedule function, and
the resulting task is saved in a dictionary.

This alleviates the need to manage the dictionary and task start/stops
in the primary class, with only simple calls to schedule() and
unschedule() being required to handle the scheduling process.