Change next dispatch time algorithm
This new algorithm is much more efficient, no longer requiring a loop
appending the interval and instead relying on simple maths to come to
the same result. This also removes the need to rely on the last dispatch
time to get the result faster, and can fully rely on the initial
dispatch time.

Manual executions should now be able to be recorded as last dispatch
times without affecting the calculation of new dispatch times.