Change event scheduler to support non repeating and late dispatches
If a non repeating event is requested to be scheduled, it first checks
to see if it is already past the timestamp before adding it to the list
of scheduled events. These are not renewed.

Late dispatches are handled so that if an event is scheduled at most 5
minutes after the previously intended dispatch time and the event had
not dispatched, it will dispatch immediately and then renew. This should
allow events to still execute if the bot had potential downtime when the
event was supposed to dispatch, but not too late that the dispatch is
completely unexpected.