Fix event not unscheduling after dispatch
Trying to cancel the task within the task wouldn't work, as an async
task cancellation requires a task switch in order to trigger. This is
manually executed through a sleep statement, doesn't matter how long,
with the applied fix using a time of 0. This makes it seem useless, but
actually isn't.