Fix sorting for single or multiple songs
Since wavelink.Playable.search always returns a list, this check should
instead check to see if the first track is part of a playlist rather
than if the object is a playlist or a playable. Playlist objects don't
seem ot be generated by this method, even when a direct playlist link is
provided.